fix: reorganize login commands into separate files
