+++ packages/cli/commands/wcp/index.js
@@ -1,5 +1,7 @@
 const login = require("./login");
+const logout = require("./logout");
+const whoami = require("./whoami");
 const project = require("./project");
 const hooks = require("./hooks");
 
-module.exports = [login(), project(), hooks()];
+module.exports = [login(), logout(), whoami(), project(), hooks()];
